[ DEV-62 ] Create seeds for classes and items

+++ src/schemas/itemsValidationSchema.ts
@@ -3,9 +3,9 @@ import { z } from 'zod';
 const itemsZodSchema = z.object({
   name: z.string(),
   description: z.string(),
-  cost: z.number().int(),
+  cost: z.number(),
   type: z.string(),
-  weight: z.number().int()
+  weight: z.number()
 });
 
 export type Item = z.infer<typeof itemsZodSchema>;
